fix update invoice account procedure

+++ src/main/kotlin/com/example/demo/entity/InvoiceUpdateAccountRequest.kt
@@ -1,3 +1,3 @@
 package com.example.demo.entity
 
-data class InvoiceUpdateAccountRequest(val oldAccountId: Long, val newAccount: Long, val invoiceId: Long)
+data class InvoiceUpdateAccountRequest(val oldAccountId: Int, val newAccount: Int, val invoiceId: Long)
